fix: move files for change directory
