+++ src/components/Icon/SpinnerIcon.tsx
