+++ src/components/Icon/TrashIcon.tsx
