+++ src/components/SearchedList.tsx
