--- src/types/todo.ts
@@ -1,6 +1,0 @@
-export interface TodoTypes {
-  title: string;
-  createdAt: string;
-  updatedAt: string;
-  id: string;
-}
